Added readme file - created populateMongo.js and script
