+++ backend/src/models/db.js
@@ -2,6 +2,10 @@ import mongoose from 'mongoose';
 
 console.log(`mongodb://${process.env.MONGODB_USERNAME}:${process.env.MONGODB_PASSWORD}@${process.env.MONGODB_URL}:${process.env.MONGODB_PORT}/${process.env.MONGODB_DATABASE}?retryWrites=true&w=majority`);
 
+// mongoose.connect(`mongodb://${process.env.MONGODB_URL}:${process.env.MONGODB_PORT}/${process.env.MONGODB_DATABASE}`);
+
+// mongoose.connect(`mongodb://${process.env.MONGODB_URL}:${process.env.MONGODB_PORT}/${process.env.MONGODB_DATABASE}`, { user: process.env.MONGODB_USERNAME, pass: process.env.MONGODB_PASSWORD, authSource: 'admin'});
+
 mongoose.connect(
     `mongodb://${process.env.MONGODB_USERNAME}:${process.env.MONGODB_PASSWORD}@${process.env.MONGODB_URL}:${process.env.MONGODB_PORT}/${process.env.MONGODB_DATABASE}?retryWrites=true&w=majority`, { authSource: process.env.MONGODB_AUTH_SOURCE }
 );
