+++ frontend/.env
@@ -1,1 +1,2 @@
+PORT=3002
 REACT_APP_BACKEND_URL=http://localhost:5000/
